requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
qrcode[pil]
Pillow
openpyxl
pandas
flask-mail
python-dotenv
requests
